Pothole Recognition/Flask App/requirements.txt
Flask==1.0.2
Keras==2.2.4
Pillow==5.2.0
plotly==4.2.1
pyparsing==2.2.0
PyYAML==3.13
requests==2.20.1
requests-file==1.4.3
requests-ftp==0.3.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
restkit==4.2.2
retrying==1.3.3
seaborn==0.9.0
simplejson==3.16.0
slackclient==1.3.1
tensorboard==1.12.0
tensorflow-gpu==1.12.0
tflearn==0.3.2
torch==0.4.1
torchvision==0.2.1
urllib3==1.24.1
webencodings==0.5.1
websocket-client==0.54.0
Werkzeug==0.14.1
